Add dotenv - Protect my secret data in .env file

+++ db/seeds.rb
@@ -11,26 +11,18 @@
 # Ingredient.create(name: "ice")
 # Ingredient.create(name: "mint leaves")
 
-require 'json'
 require 'open-uri'
 
-Ingredient.destroy_all
-Cocktail.destroy_all
-Dose.destroy_all
+puts "Destroy ingredients"
+Ingredient.destroy_all if Rails.env.development?
 
-url = 'http://www.thecocktaildb.com/api/json/v1/1/list.php?i=list'
-ingredients_serialized = open(url).read
-ingredients = JSON.parse(ingredients_serialized)
+puts "Destroy Cocktails"
+Cocktail.destroy_all if Rails.env.development?
 
+puts "Create ingredients"
+url = "https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list"
+ingredients = JSON.parse(open(url).read)
 ingredients["drinks"].each do |ingredient|
-  Ingredient.create(name: ingredient["strIngredient1"])
-
-end
-
-url = 'http://www.thecocktaildb.com/api/json/v1/1/filter.php?a=Alcoholic'
-cocktails_serialized = open(url).read
-cocktails = JSON.parse(cocktails_serialized)
-
-cocktails["drinks"].each do |cocktail|
-  Cocktail.create(name: cocktail["strDrink"])
+  i = Ingredient.create(name: ingredient["strIngredient1"])
+  puts "create #{i.name}"
 end
